deployment/requirements.txt
streamlit==1.44.0
pandas==2.2.2
numpy==1.26.4
cloudpickle==3.1.1
scikit-learn==1.6.1
seaborn==0.13.2
matplotlib==3.9.4
plotly==5.24.1
pillow